fix(analytics): :label: change TotalClicks type and add Firestore tags to ClickLog fields
change TotalClicks type from int to int64 for accurate click count and add Firestore tags to ClickLog fields for proper serialization

+++ internal/dto/analytics_dto.go
@@ -4,6 +4,6 @@ import "github.com/mfmahendr/url-shortener-backend/internal/models"
 
 type AnalyticsDTO struct {
 	ShortID     string            `json:"short_id"`
-	TotalClicks int               `json:"total_clicks"`
+	TotalClicks int64             `json:"total_clicks"`
 	Clicks      []models.ClickLog `json:"clicks"`
 }
